#171 refactored UserTDG, one of the admins can not be deleted

+++ class/TDG/UserTDG.js
@@ -2,11 +2,11 @@ const mysql = require('mysql');
 const express = require('express');
 var app = express();
 const bodyparser = require('body-parser');
-const encrypt = require("js-sha512"); 
+const encrypt = require("js-sha512");
 
 class userTDG {
 	//constructor is used to create a connection to the database
-	constructor(){
+	constructor() {
 		this.mysqlConnection = mysql.createPool({
 			host: '192.185.72.57',
 			user: 'arti17co_soen343',
@@ -14,38 +14,38 @@ class userTDG {
 			database: 'arti17co_soen343',
 			multipleStatements: true,
 		});
-		this.runQuery = function (queryBuild){
+		this.runQuery = function (queryBuild) {
 			let conn = this.mysqlConnection;
-			queryBuild(conn,function(type){
-				console.log("Completed query "+type+" \n");
+			queryBuild(conn, function (type) {
+				console.log("Completed query " + type + " \n");
 			});
 		};
 	}
 
-	login (email, password,callback){
-		let sql= "SELECT UserId,FirstName,LastName, Address, email, phone, type FROM User WHERE email ='"+email+"' AND password = +'"+encrypt.sha512(password)+"'";
-		this.runQuery(function(conn,completedQuery){
+	login(email, password, callback) {
+		let sql = "SELECT UserId,FirstName,LastName, Address, email, phone, type FROM User WHERE email ='" + email + "' AND password = +'" + encrypt.sha512(password) + "'";
+		this.runQuery(function (conn, completedQuery) {
 			conn.query(sql, (err, rows, fields) => {
-				if (!err){
-					if(rows.length==1){
+				if (!err) {
+					if (rows.length == 1) {
 						let msg = {};
-						msg.success="true";
-						msg.message="no message";
-						msg.data=rows[0];
+						msg.success = "true";
+						msg.message = "no message";
+						msg.data = rows[0];
 						callback(msg);
 					}
-					else{
+					else {
 						let msg = {};
-						msg.success="false";
-						msg.message="incorrect username or password";
+						msg.success = "false";
+						msg.message = "incorrect username or password";
 						callback(msg);
 					}
-				}	 
-				else{
+				}
+				else {
 					console.log(err);
 					let msg = {};
-					msg.success="false";
-					msg.message="internal error";
+					msg.success = "false";
+					msg.message = "internal error";
 					console.log(err);
 					callback(msg);
 				}
@@ -53,23 +53,23 @@ class userTDG {
 			})
 		})
 	}
-	
-	viewAllUsers(callback){
-		let sql = 'SELECT UserId, FirstName, LastName, Address, email, phone, type FROM User' ; 
-		this.runQuery(function(conn,completedQuery){
+
+	viewAllUsers(callback) {
+		let sql = 'SELECT UserId, FirstName, LastName, Address, email, phone, type FROM User';
+		this.runQuery(function (conn, completedQuery) {
 			conn.query(sql, (err, rows, fields) => {
-				if (!err){
+				if (!err) {
 					let msg = {};
-					msg.success="true";
-					msg.message="no message";
-					msg.data=rows;
+					msg.success = "true";
+					msg.message = "no message";
+					msg.data = rows;
 					callback(msg);
-				}	 
-				else{
+				}
+				else {
 					console.log(err);
 					let msg = {};
-					msg.success="false";
-					msg.message=err.sqlMessage;;
+					msg.success = "false";
+					msg.message = err.sqlMessage;;
 					callback(msg);
 				}
 				completedQuery("[UserTDG] viewAllUsers()");
@@ -77,23 +77,23 @@ class userTDG {
 		})
 	}
 
-	addUser(user, callback){
-		let sql = "INSERT INTO User (FirstName, LastName, Address, email, phone, type, password)"+
-		"VALUES('"+user.FirstName+"', '"+user.LastName+"', '"+user.Address+"', '"+user.email+"', '"+user.phone+"', '"+user.type+"', '"+encrypt.sha512(user.password)+"');"
-		this.runQuery(function(conn,completedQuery){
+	addUser(user, callback) {
+		let sql = "INSERT INTO User (FirstName, LastName, Address, email, phone, type, password)" +
+			"VALUES('" + user.FirstName + "', '" + user.LastName + "', '" + user.Address + "', '" + user.email + "', '" + user.phone + "', '" + user.type + "', '" + encrypt.sha512(user.password) + "');"
+		this.runQuery(function (conn, completedQuery) {
 			conn.query(sql, (err, rows, fields) => {
-				if (!err){
+				if (!err) {
 					let msg = {};
-					msg.success="true";
-					msg.message="no message";
-					msg.data=rows;
+					msg.success = "true";
+					msg.message = "no message";
+					msg.data = rows;
 					callback(msg);
-				}	 
-				else{
+				}
+				else {
 					console.log(err);
 					let msg = {};
-					msg.success="false";
-					msg.message=err.sqlMessage;;
+					msg.success = "false";
+					msg.message = err.sqlMessage;;
 					callback(msg);
 				}
 				completedQuery("[UserTDG] addUser()");
@@ -101,25 +101,25 @@ class userTDG {
 		})
 	}
 
-	modifyUser(user, callback){
-		let sql = "UPDATE User "+
-		"SET Firstname = '"+user.FirstName+"', LastName = '"+user.LastName+"', Address = '"+user.Address+"', email = '"+user.email+"', phone = '"+user.phone+"', type = '"+user.type+"', password = '"+encrypt.sha512(user.password)+"'"+
-		"WHERE UserId = '"+user.UserId+"';"
-		
-		this.runQuery(function(conn,completedQuery){
+	modifyUser(user, callback) {
+		let sql = "UPDATE User " +
+			"SET Firstname = '" + user.FirstName + "', LastName = '" + user.LastName + "', Address = '" + user.Address + "', email = '" + user.email + "', phone = '" + user.phone + "', type = '" + user.type + "', password = '" + encrypt.sha512(user.password) + "'" +
+			"WHERE UserId = '" + user.UserId + "';"
+
+		this.runQuery(function (conn, completedQuery) {
 			conn.query(sql, (err, rows, fields) => {
-				if (!err){
+				if (!err) {
 					let msg = {};
-					msg.success="true";
-					msg.message="no message";
-					msg.data=rows;
+					msg.success = "true";
+					msg.message = "no message";
+					msg.data = rows;
 					callback(msg);
-				}	 
-				else{
+				}
+				else {
 					console.log(err);
 					let msg = {};
-					msg.success="false";
-					msg.message=err.sqlMessage;;
+					msg.success = "false";
+					msg.message = err.sqlMessage;;
 					callback(msg);
 				}
 				completedQuery("[UserTDG] modifyUser()");
@@ -127,22 +127,29 @@ class userTDG {
 		})
 	}
 
-	deleteUser(id,callback){
-		let sql=  '  DELETE FROM User where UserId = '+id+';  ' ;
-		this.runQuery(function(conn,completedQuery){
+	deleteUser(id, callback) {
+		if (id == 1) {
+			let msg = {};
+			msg.alert = "User can not be deleted, he is the Sensei";
+			callback(msg);
+			return msg;
+		}
+
+		let sql = '  DELETE FROM User where UserId = ' + id + ';  ';
+		this.runQuery(function (conn, completedQuery) {
 			conn.query(sql, (err, rows, fields) => {
-				if (!err){
+				if (!err) {
 					let msg = {};
-					msg.success="true";
-					msg.message="no message";
-					msg.data=rows;
+					msg.success = "true";
+					msg.message = "no message";
+					msg.data = rows;
 					callback(msg);
-				}	 
-				else{
+				}
+				else {
 					console.log(err);
 					let msg = {};
-					msg.success="false";
-					msg.message=err.sqlMessage;;
+					msg.success = "false";
+					msg.message = err.sqlMessage;;
 					callback(msg);
 				}
 				completedQuery("[UserTDG] deleteUser()");
